requirements.txt
# ──────────────────────────────────────────────
# SafeSpace AI — Python Dependencies
# ──────────────────────────────────────────────

# Web Framework & Server
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.20
pydantic==2.10.4
python-dotenv==1.0.1

# AI Engine
groq==1.1.1

# Privacy & Anonymization
presidio-analyzer==2.2.356
presidio-anonymizer==2.2.356
spacy==3.8.4

# Research Dashboard
streamlit==1.41.1
pandas==2.2.3
plotly==5.24.1

# Utilities
httpx==0.28.1

# Production Database (optional — only needed for Render/cloud deploy)
# psycopg2-binary==2.9.11